move into model/commitqueue
