+++ model/lifecycle.go
@@ -11,7 +11,6 @@ import (
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/db"
 	"github.com/evergreen-ci/evergreen/model/build"
-	"github.com/evergreen-ci/evergreen/model/commitqueue"
 	"github.com/evergreen-ci/evergreen/model/distro"
 	"github.com/evergreen-ci/evergreen/model/event"
 	"github.com/evergreen-ci/evergreen/model/patch"
@@ -21,7 +20,7 @@ import (
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/message"
 	"github.com/pkg/errors"
-	mgo "gopkg.in/mgo.v2"
+	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/bson"
 )
 
@@ -175,83 +174,6 @@ func AbortVersion(versionId, caller string) error {
 	if len(ids) > 0 {
 		event.LogManyTaskAbortRequests(ids, caller)
 	}
-	return nil
-}
-
-func RemoveCommitQueueItem(projectId, item string) (bool, error) {
-	cq, err := commitqueue.FindOneId(projectId)
-	if err != nil {
-		return false, errors.Wrapf(err, "can't get commit queue for id '%s'", projectId)
-	}
-
-	head := cq.Next()
-	removed, err := cq.Remove(item)
-	if err != nil {
-		return removed, errors.Wrapf(err, "can't remove item '%s' from queue '%s'", item, projectId)
-	}
-
-	if removed && head.Issue == item {
-		if err = preventMergeForItem(projectId, head); err != nil {
-			return removed, errors.Wrapf(err, "can't prevent merge for item '%s' on queue '%s'", item, projectId)
-		}
-	}
-	return removed, nil
-}
-
-func preventMergeForItem(projectID string, item *commitqueue.CommitQueueItem) error {
-	projectRef, err := FindOneProjectRef(projectID)
-	if err != nil {
-		return errors.Wrapf(err, "can't find projectRef for '%s'", projectID)
-	}
-	if projectRef == nil {
-		return errors.Errorf("can't find project ref for '%s'", projectID)
-	}
-
-	if projectRef.CommitQueue.PatchType == commitqueue.PRPatchType && item.Version != "" {
-		if err = clearVersionPatchSubscriber(item.Version, event.GithubMergeSubscriberType); err != nil {
-			return errors.Wrap(err, "can't clear subscriptions")
-		}
-	}
-
-	if projectRef.CommitQueue.PatchType == commitqueue.CLIPatchType {
-		version, err := VersionFindOneId(item.Issue)
-		if err != nil {
-			return errors.Wrapf(err, "can't find patch '%s'", item.Issue)
-		}
-		if version != nil {
-			if err = clearVersionPatchSubscriber(version.Id, event.CommitQueueDequeueSubscriberType); err != nil {
-				return errors.Wrap(err, "can't clear subscriptions")
-			}
-
-			// Blacklist the merge task
-			mergeTask, err := task.FindMergeTaskForVersion(version.Id)
-			if err != nil {
-				return errors.Wrapf(err, "can't find merge task for '%s'", version.Id)
-			}
-			err = mergeTask.SetPriority(-1, evergreen.User)
-			if err != nil {
-				return errors.Wrap(err, "can't blacklist merge task")
-			}
-		}
-	}
-
-	return nil
-}
-
-func clearVersionPatchSubscriber(versionID, subscriberType string) error {
-	subscriptions, err := event.FindSubscriptions(event.ResourceTypePatch, []event.Selector{{Type: event.SelectorID, Data: versionID}})
-	if err != nil {
-		return errors.Wrapf(err, "can't find subscription to patch '%s'", versionID)
-	}
-	for _, subscription := range subscriptions {
-		if subscription.Subscriber.Type == subscriberType {
-			err = event.RemoveSubscription(subscription.ID)
-			if err != nil {
-				return errors.Wrapf(err, "can't remove subscription for '%s', type '%s'", versionID, subscriberType)
-			}
-		}
-	}
-
 	return nil
 }
